+++ include/real/real.hpp
@@ -6,6 +6,7 @@
 #include <vector>
 #include <regex>
 #include <initializer_list>
+#include <sstream>
 #include <utility>
 #include <memory> // shared_ptr
 #include <variant>
@@ -110,7 +111,7 @@ namespace boost {
                     integer_part = integer_part.substr(1);
                 }
                 integer_part = regex_replace(integer_part, std::regex("(0?+)([[:digit:]]?+)"), "$2");
-                int i = decimal_part.length() - 1;
+                size_t i = decimal_part.length() - 1;
                 while (decimal_part[i] == '0' && i >= 0) {
                     --i;
                 }
